Add files via upload
Just some examples, this makes it easier to test the above graphs. You can just copy and paste the numbers when the program asks for each graph. A picture is shown for the first two to help with visualization. 
if the graph is  
0 1
1 0
the first 0 1 is saying that point a is connected to point b only hence a 1 in the second position representing b.
and the second 1 0 is saying that point b is connected to point a only hence a 1 in the first position representing a.